moved navbar to separate file
